Fixed all warnings

+++ source/montecarlo.h
@@ -10,5 +10,5 @@
 #include <iostream>
 
 void   metropolis(const std::vector<Node> &Site, struct MC_parameters &MCp, struct H_parameters &Hp, double my_beta);
-double local_HPsi(std::array<O2, NC> &Psi, int ix, int iy, struct H_parameters &Hp, const std::vector<Node> &Site);
-double local_HA(double A, int ix, int iy, int alpha, struct H_parameters &Hp, const std::vector<Node> &Site);
+double local_HPsi(std::array<O2, NC> &Psi, size_t ix, size_t iy, struct H_parameters &Hp, const std::vector<Node> &Site);
+double local_HA(double A, size_t ix, size_t iy, size_t alpha, struct H_parameters &Hp, const std::vector<Node> &Site);
